+++ src/helper/parseQuery.ts
@@ -1,6 +1,7 @@
 import { FilterQuery, isValidObjectId, Types } from "mongoose";
 import { Book } from "../Book";
 import { isValidTransactionKey, ITransaction } from "../models/transactions";
+import { parseAccountField } from "./parseAccountField";
 import { parseDateField } from "./parseDateField";
 
 export type IParseQuery = {
@@ -25,7 +26,7 @@ export function parseQuery(
   query: IParseQuery & IPaginationQuery,
   book: Pick<Book, "name">
 ): FilterQuery<ITransaction> {
-  let i, il, j, jl;
+  let i, il;
 
   const { approved, account, start_date, end_date, ...extra } = query;
 
@@ -35,30 +36,8 @@ export function parseQuery(
   const filterQuery: FilterQuery<ITransaction> = {
     book: book.name,
     approved: approved !== false,
+    ...parseAccountField(account),
   };
-
-  if (account) {
-    if (Array.isArray(account) && account.length === 1) {
-      const accounts = account[0].split(":");
-      for (i = 0, il = accounts.length; i < il; i++) {
-        filterQuery[`account_path.${i}`] = accounts[i];
-      }
-    } else if (Array.isArray(account)) {
-      filterQuery["$or"] = new Array(account.length);
-      for (i = 0, il = account.length; i < il; i++) {
-        const accounts = account[i].split(":");
-        filterQuery["$or"][i] = {};
-        for (j = 0, jl = accounts.length; j < jl; j++) {
-          filterQuery["$or"][i][`account_path.${j}`] = accounts[j];
-        }
-      }
-    } else {
-      const accounts = account.split(":");
-      for (i = 0, il = accounts.length; i < il; i++) {
-        filterQuery[`account_path.${i}`] = accounts[i];
-      }
-    }
-  }
 
   if (start_date || end_date) {
     filterQuery["datetime"] = {};
